2023.10.23

+++ src/main/java/com/example/structure/body/dto/request/BodyRequest.java
@@ -8,10 +8,10 @@ import javax.validation.constraints.NotBlank;
 @RequiredArgsConstructor
 public class BodyRequest {
     @NotBlank(message="몸무게를 입력해주세요")
-    private final Long weight;
+    private final float weight;
     @NotBlank(message="체지방률 입력해주세요")
-    private final Long fat;
+    private final float fat;
     @NotBlank(message="골격근량을 입력해주세요")
-    private final Long muscle;
+    private final float muscle;
 
 }
